refactor: TaskCard and related components to separate file
